+++ languages/ada/queries/highlights.scm
@@ -26,7 +26,6 @@
    "interface"
    "is"
    "limited"
-   "null"
    "of"
    "others"
    "out"
@@ -41,10 +40,13 @@
    "when"
 ] @keyword
 [
+   "null"
+] @constant.builtin
+[
    "aliased"
    "constant"
    "renames"
-] @storageclass
+] @keyword.storage
 [
    "mod"
    "new"
@@ -52,11 +54,11 @@
    "record"
    "subtype"
    "type"
-] @keyword.type
+] @type.builtin
 [
    "with"
    "use"
-] @include
+] @keyword.control.import
 [
    "body"
    "function"
@@ -71,7 +73,7 @@
    "not"
    "or"
    "xor"
-] @keyword.operator
+] @operator
 [
    "while"
    "loop"
@@ -79,10 +81,10 @@
    "parallel"
    "reverse"
    "some"
-] @repeat
+] @kewyord.control.repeat
 [
    "return"
-] @keyword.return
+] @keyword.control.return
 [
    "case"
    "if"
@@ -90,70 +92,70 @@
    "then"
    "elsif"
    "select"
-] @conditional
+] @keyword.control.conditional
 [
    "exception"
    "raise"
-] @exception
-(comment) @comment @spell
-(string_literal) @string
+] @keyword.control.exception
+(comment)         @comment
+(string_literal)  @string
 (character_literal) @string
-(numeric_literal) @number
+(numeric_literal) @constant.numeric
 
 ;; Highlight the name of subprograms
-(procedure_specification name: (_) @function)
-(function_specification name: (_) @function)
-(package_declaration name: (_) @function)
-(package_body name: (_) @function)
-(generic_instantiation name: (_) @function)
-(entry_declaration . (identifier) @function)
+(procedure_specification name: (_) @function.builtin)
+(function_specification name: (_) @function.builtin)
+(package_declaration name: (_) @function.builtin)
+(package_body name: (_) @function.builtin)
+(generic_instantiation name: (_) @function.builtin)
+(entry_declaration . (identifier) @function.builtin)
 
 ;; Some keywords should take different categories depending on the context
-(use_clause "use"  @include "type" @include)
-(with_clause "private" @include)
-(with_clause "limited" @include)
+(use_clause "use"  @keyword.control.import "type" @keyword.control.import)
+(with_clause "private" @keyword.control.import)
+(with_clause "limited" @keyword.control.import)
 (use_clause (_) @namespace)
 (with_clause (_) @namespace)
 
-(loop_statement "end" @keyword.repeat)
-(if_statement "end" @conditional)
-(loop_parameter_specification "in" @keyword.repeat)
-(loop_parameter_specification "in" @keyword.repeat)
-(iterator_specification ["in" "of"] @keyword.repeat)
-(range_attribute_designator "range" @keyword.repeat)
+(loop_statement "end" @keyword.control.repeat)
+(if_statement "end" @keyword.control.conditional)
+(loop_parameter_specification "in" @keyword.control.repeat)
+(loop_parameter_specification "in" @keyword.control.repeat)
+(iterator_specification ["in" "of"] @keyword.control.repeat)
+(range_attribute_designator "range" @keyword.control.repeat)
 
-(raise_statement "with" @exception)
+(raise_statement "with" @keyword.control.exception)
 
-(gnatprep_declarative_if_statement)  @preproc
-(gnatprep_if_statement)              @preproc
-(gnatprep_identifier)                @preproc
+(gnatprep_declarative_if_statement)  @keyword.directive
+(gnatprep_if_statement)              @keyword.directive
+(gnatprep_identifier)                @keyword.directive
 
 (subprogram_declaration "is" @keyword.function "abstract"  @keyword.function)
 (aspect_specification "with" @keyword.function)
 
-(full_type_declaration "is" @keyword.type)
-(subtype_declaration "is" @keyword.type)
-(record_definition "end" @keyword.type)
-(full_type_declaration (_ "access" @keyword.type))
-(array_type_definition "array" @keyword.type "of" @keyword.type)
-(access_to_object_definition "access" @keyword.type)
-(access_to_object_definition "access" @keyword.type
+(full_type_declaration "is" @type.builtin)
+(subtype_declaration "is" @type.builtin)
+(record_definition "end" @type.builtin)
+(full_type_declaration (_ "access" @type.builtin))
+(array_type_definition "array" @type.builtin "of" @type.builtin)
+(access_to_object_definition "access" @type.builtin)
+(access_to_object_definition "access" @type.builtin
    [
-      (general_access_modifier "constant" @keyword.type)
-      (general_access_modifier "all" @keyword.type)
+      (general_access_modifier "constant" @type.builtin)
+      (general_access_modifier "all" @type.builtin)
    ]
 )
-(range_constraint "range" @keyword.type)
-(signed_integer_type_definition "range" @keyword.type)
-(index_subtype_definition "range" @keyword.type)
-(record_type_definition "abstract" @keyword.type)
-(record_type_definition "tagged" @keyword.type)
-(record_type_definition "limited" @keyword.type)
-(record_type_definition (record_definition "null" @keyword.type))
-(private_type_declaration "is" @keyword.type "private" @keyword.type)
-(private_type_declaration "tagged" @keyword.type)
-(private_type_declaration "limited" @keyword.type)
-(task_type_declaration "task" @keyword.type "is" @keyword.type)
+(range_constraint "range" @type.builtin)
+(signed_integer_type_definition "range" @type.builtin)
+(index_subtype_definition "range" @type.builtin)
+(record_type_definition "abstract" @type.builtin)
+(record_type_definition "tagged" @type.builtin)
+(record_type_definition "limited" @type.builtin)
+(record_type_definition (record_definition "null" @type.builtin))
+(private_type_declaration "is" @type.builtin "private" @type.builtin)
+(private_type_declaration "tagged" @type.builtin)
+(private_type_declaration "limited" @type.builtin)
+(task_type_declaration "task" @type.builtin "is" @type.builtin)
 
 ;; Gray the body of expression functions
 (expression_function_declaration
@@ -164,33 +166,11 @@
 (subprogram_declaration (aspect_specification) @attribute)
 
 ;; Highlight full subprogram specifications
-;(subprogram_body
-;    [
-;       (procedure_specification)
-;       (function_specification)
-;    ] @function.spec
-;)
+; (subprogram_body
+;     [
+;        (procedure_specification)
+;        (function_specification)
+;     ] @function.builtin.spec
+; )
 
-((comment) @comment.documentation
-  . [
-      (entry_declaration)
-      (subprogram_declaration)
-      (parameter_specification)
-    ])
-
-(compilation_unit 
-  . (comment) @comment.documentation)
-
-(component_list
-  (component_declaration)
-    . (comment) @comment.documentation)
-
-(enumeration_type_definition 
-  (identifier)
-  . (comment) @comment.documentation)
-
-;; Highlight errors in red. This is not very useful in practice, as text will
-;; be highlighted as user types, and the error could be elsewhere in the code.
-;; This also requires defining    :hi @error guifg=Red    for instance.
-(ERROR) @error